# storage/engine: micro-optimizations for MVCCMerge