Added a benchmark which mimics the periodic writing of timeseries
values.

```
name                              old time/op  new time/op  delta
MVCCBatchTimeSeries282_RocksDB-8   323µs ± 2%   297µs ± 3%  -7.89%  (p=0.000 n=20+20)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/9862)
<!-- Reviewable:end -->